Combined retry and repeat in one operation. Took the idea further to also allow the notification function control the timing of the retry/repeat.

+++ rxjava-core/src/main/java/rx/internal/operators/OperatorTake.java
@@ -29,9 +29,9 @@ import rx.Subscriber;
  */
 public final class OperatorTake<T> implements Operator<T, T> {
 
-    final int limit;
+    final long limit;
 
-    public OperatorTake(int limit) {
+    public OperatorTake(long limit) {
         this.limit = limit;
     }
